@@ -1,4 +1,3 @@
-import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { ThemeProvider } from './contexts/ThemeContext';
 import { AuthProvider } from './contexts/AuthContext';
@@ -7,23 +6,41 @@ import Products from './pages/Products';
 import Login from './pages/Login';
 import Signup from './pages/Signup';
 import Orders from './pages/Orders';
+import AdminLayout from './pages/admin/AdminLayout';
+import AdminDashboard from './pages/admin/AdminDashboard';
+import ManageProducts from './pages/admin/ManageProducts';
+import ManageOrders from './pages/admin/ManageOrders';
+import ManageUsers from './pages/admin/ManageUsers';
 
 function App() {
   return (
     <Router>
       <ThemeProvider>
         <AuthProvider>
-          <div className="min-h-screen bg-gray-50 dark:bg-navy-900 transition-colors duration-300">
-            <Header />
-            <main>
-              <Routes>
-                <Route path="/" element={<Products />} />
-                <Route path="/login" element={<Login />} />
-                <Route path="/signup" element={<Signup />} />
-                <Route path="/orders" element={<Orders />} />
-              </Routes>
-            </main>
-          </div>
+          <Routes>
+            {/* Admin Routes */}
+            <Route path="/admin" element={<AdminLayout />}>
+              <Route index element={<AdminDashboard />} />
+              <Route path="products" element={<ManageProducts />} />
+              <Route path="orders" element={<ManageOrders />} />
+              <Route path="users" element={<ManageUsers />} />
+            </Route>
+
+            {/* Public Routes */}
+            <Route path="*" element={
+              <div className="min-h-screen bg-gray-50 dark:bg-navy-900 transition-colors duration-300">
+                <Header />
+                <main>
+                  <Routes>
+                    <Route path="/" element={<Products />} />
+                    <Route path="/login" element={<Login />} />
+                    <Route path="/signup" element={<Signup />} />
+                    <Route path="/orders" element={<Orders />} />
+                  </Routes>
+                </main>
+              </div>
+            } />
+          </Routes>
         </AuthProvider>
       </ThemeProvider>
     </Router>
